Add support for _Locales.
Close #4

+++ src/file_util.cpp
@@ -405,9 +405,9 @@ void CloseMPQArchive(void* MPQ)
   SFileCloseArchive(MPQ);
 }
 
-optional<uint32_t> GetMPQFileSize(void* MPQ, const char* packedFileName, const uint32_t locale)
+optional<uint32_t> GetMPQFileSize(void* MPQ, const char* packedFileName, const uint16_t locale)
 {
-  SFileSetLocale(locale);
+  SFileSetLocale((uint32_t)locale);
   void* subFile = nullptr;
   optional<uint32_t> fileLength;
   if (SFileOpenFileEx(MPQ, packedFileName, 0, &subFile)) {
@@ -417,7 +417,7 @@ optional<uint32_t> GetMPQFileSize(void* MPQ, const char* packedFileName, const u
   return fileLength;
 }
 
-bool ExtractMPQFile(void* MPQ, const char* packedFileName, const filesystem::path& outPath, const uint32_t locale)
+bool ExtractMPQFile(void* MPQ, const char* packedFileName, const filesystem::path& outPath, const uint16_t locale)
 {
   vector<uint8_t> container;
   ReadMPQFile(MPQ, packedFileName, container, locale);
